modfied

+++ src/app/api/link/delete/route.ts
@@ -1,25 +1,28 @@
 import { NextResponse } from "next/server";
 
-import { collection, doc, deleteField, updateDoc } from 'firebase/firestore';
-import { db } from "@/database/firebaseConfig";
+import { collection, doc, getDoc, deleteField, updateDoc } from 'firebase/firestore';
+import { db } from "@/database/firebase";
 
 export async function DELETE(req : Request) 
 {
-    try {
-        const body = await req.json();
+    const { parentSectionID, linkID } = await req.json();
 
+    try 
+    {
         const collectionRef = collection(db, 'document');
-        const documentRef = doc(db, collectionRef.id, body.sectionId)
-        
-        const linkID = body.linkId;
+        const documentReference = doc(db, collectionRef.id, parentSectionID)
+        const documentSnapshot = await getDoc(documentReference);
 
-        await updateDoc(documentRef, {
-            [linkID] : deleteField()
-        })
+        if(documentSnapshot.exists()) {
+            const response = await updateDoc(documentReference, {
+                [linkID] : deleteField()
+            })
+            return NextResponse.json({ status : 200, message : `field deleted`, code : response });
+        }
+        return NextResponse.json({ status : 404, message : 'parent document was not found', code : '' });
 
-        return NextResponse.json({ status : 'field deleted', code : 200 });
     } 
-    catch (error) {
-        return NextResponse.json({ status : `internal server error ${error}`, code : 500 });
+    catch (error : any) {
+        return NextResponse.json({ status : 500, message : `internal server error`, code : error });
     }
 }
